+++ README.md
@@ -5,15 +5,13 @@ REST API for BitShares blockchain gateway
 ```shell script
 git clone https://github.com/fincubator/gateways_api
 cd gateways_api/
-cp config/config.yml.example config/config.yml
-cp docker-compose.yml.example docker-compose.yml
-cp src/alembic.ini.example src/alembic.ini
+cp config/config.yml.example config/config.yml && cp docker-compose.yml.example docker-compose.yml && cp migrations/alembic.ini.example migrations/alembic.ini
 ```
 
 if you want to use your own PostgreSQL connection data, you need to change:
 * `config/config.yml`
 * *postgres:environment* in `docker-compose.yml`
-* line 38 in `src/alembic.ini`
+* line 38 in `alembic.ini`
 
 ### Build & Run via Docker
 
@@ -27,7 +25,7 @@ $ sudo docker exec -it gw_api_container_ID bash
 Run migration and unittest inside *gw_api* container. 
 Unitests fill the database with test data, don't forget to remove it on production deploy
 ```shell script
-root@gw_api_container_ID:/app# cd src/
+root@gw_api_container_ID:/app# cd migrations/
 root@gw_api_container_ID:/app# alembic upgrade head
 root@gw_api_container_ID:/app# cd ..
 root@gw_api_container_ID:/app# python -m unittest
