+++ requirements.txt
@@ -9,4 +9,7 @@ aiopg~=1.0.0
 alembic~=1.4.2
 
 PyYAML~=5.3.1
+marshmallow
+
 gunicorn
+
