+++ tests/test_database.py
@@ -1,7 +1,7 @@
 from aiohttp.test_utils import AioHTTPTestCase, unittest_run_loop
 
 from app import app_factory
-from src.database.db_utils import get_all_assets, add_asset
+from db_utils import get_all_assets, add_asset
 
 from tests.fixtures import test_assets_data
